// web/admin/src/types/api/v1/models/CreateExperienceRequest.ts
export interface CreateExperienceRequest {
    /**
     * 営業終了時間
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    businessCloseTime: string;
    /**
     * 営業開始時間
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    businessOpenTime: string;
    /**
     * コーディネータID
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    coordinatorId: string;
    /**
     * 説明
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    description: string;
    /**
     * アクセス方法
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    direction: string;
    /**
     * 体験時間(分)
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    duration: number;
    /**
     * 募集終了日時
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    endAt: number;
    /**
     * 体験種別ID
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    experienceTypeId: string;
    /**
     * 開催場所(住所1)
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    hostAddressLine1: string;
    /**
     * 開催場所(住所2)
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    hostAddressLine2: string;
    /**
     * 開催場所(市区町村)
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    hostCity: string;
    /**
     * 開催場所(郵便番号)
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    hostPostalCode: string;
    /**
     * 開催場所(都道府県コード)
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    hostPrefectureCode: number;
    /**
     * メディア一覧
     * @type {Array<CreateExperienceMedia>}
     * @memberof CreateExperienceRequest
     */
    media: Array<CreateExperienceMedia>;
    /**
     * 大人料金
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    priceAdult: number;
    /**
     * 小学生料金
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    priceElementarySchool: number;
    /**
     * 中学生料金
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    priceJuniorHighSchool: number;
    /**
     * 幼児料金
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    pricePreschool: number;
    /**
     * シニア料金
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    priceSenior: number;
    /**
     * 生産者ID
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    producerId: string;
    /**
     * 紹介動画URL
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    promotionVideoUrl: string;
    /**
     * 公開設定
     * @type {boolean}
     * @memberof CreateExperienceRequest
     */
    _public: boolean;
    /**
     * おすすめポイント1
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    recommendedPoint1: string;
    /**
     * おすすめポイント2
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    recommendedPoint2: string;
    /**
     * おすすめポイント3
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    recommendedPoint3: string;
    /**
     * 定員オーバーフラグ
     * @type {boolean}
     * @memberof CreateExperienceRequest
     */
    soldOut: boolean;
    /**
     * 募集開始日時
     * @type {number}
     * @memberof CreateExperienceRequest
     */
    startAt: number;
    /**
     * 体験名
     * @type {string}
     * @memberof CreateExperienceRequest
     */
    title: string;
}