// web/admin/src/types/api/v1/models/Experience.ts
export interface Experience {
    /**
     * 営業終了時間
     * @type {string}
     * @memberof Experience
     */
    businessCloseTime: string;
    /**
     * 営業開始時間
     * @type {string}
     * @memberof Experience
     */
    businessOpenTime: string;
    /**
     * コーディネータID
     * @type {string}
     * @memberof Experience
     */
    coordinatorId: string;
    /**
     * 作成日時
     * @type {number}
     * @memberof Experience
     */
    createdAt: number;
    /**
     * 説明
     * @type {string}
     * @memberof Experience
     */
    description: string;
    /**
     * アクセス方法
     * @type {string}
     * @memberof Experience
     */
    direction: string;
    /**
     * 体験時間(分)
     * @type {number}
     * @memberof Experience
     */
    duration: number;
    /**
     * 募集終了日時
     * @type {number}
     * @memberof Experience
     */
    endAt: number;
    /**
     * 体験種別ID
     * @type {string}
     * @memberof Experience
     */
    experienceTypeId: string;
    /**
     * 開催場所(住所1)
     * @type {string}
     * @memberof Experience
     */
    hostAddressLine1: string;
    /**
     * 開催場所(住所2)
     * @type {string}
     * @memberof Experience
     */
    hostAddressLine2: string;
    /**
     * 開催場所(市区町村)
     * @type {string}
     * @memberof Experience
     */
    hostCity: string;
    /**
     * 開催場所(郵便番号)
     * @type {string}
     * @memberof Experience
     */
    hostPostalCode: string;
    /**
     * 開催場所(都道府県コード)
     * @type {number}
     * @memberof Experience
     */
    hostPrefectureCode: number;
    /**
     * 体験ID
     * @type {string}
     * @memberof Experience
     */
    id: string;
    /**
     * メディア一覧
     * @type {Array<ExperienceMedia>}
     * @memberof Experience
     */
    media: Array<ExperienceMedia>;
    /**
     * 大人料金
     * @type {number}
     * @memberof Experience
     */
    priceAdult: number;
    /**
     * 小学生料金
     * @type {number}
     * @memberof Experience
     */
    priceElementarySchool: number;
    /**
     * 中学生料金
     * @type {number}
     * @memberof Experience
     */
    priceJuniorHighSchool: number;
    /**
     * 幼児料金
     * @type {number}
     * @memberof Experience
     */
    pricePreschool: number;
    /**
     * シニア料金
     * @type {number}
     * @memberof Experience
     */
    priceSenior: number;
    /**
     * プロデューサーID
     * @type {string}
     * @memberof Experience
     */
    producerId: string;
    /**
     * 紹介動画URL
     * @type {string}
     * @memberof Experience
     */
    promotionVideoUrl: string;
    /**
     * 公開設定
     * @type {boolean}
     * @memberof Experience
     */
    _public: boolean;
    /**
     * おすすめポイント1
     * @type {string}
     * @memberof Experience
     */
    recommendedPoint1: string;
    /**
     * おすすめポイント2
     * @type {string}
     * @memberof Experience
     */
    recommendedPoint2: string;
    /**
     * おすすめポイント3
     * @type {string}
     * @memberof Experience
     */
    recommendedPoint3: string;
    /**
     * 定員オーバー設定
     * @type {boolean}
     * @memberof Experience
     */
    soldOut: boolean;
    /**
     * 募集開始日時
     * @type {number}
     * @memberof Experience
     */
    startAt: number;
    /**
     * 
     * @type {ExperienceStatus}
     * @memberof Experience
     */
    status: ExperienceStatus;
    /**
     * タイトル
     * @type {string}
     * @memberof Experience
     */
    title: string;
    /**
     * 更新日時
     * @type {number}
     * @memberof Experience
     */
    updatedAt: number;
}