// web/admin/src/types/api/v1/models/ExperiencesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 管理者向け
 * 管理者向けのふるマルAPIです。
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Experience } from './Experience';
import {
    ExperienceFromJSON,
    ExperienceFromJSONTyped,
    ExperienceToJSON,
    ExperienceToJSONTyped,
} from './Experience';
import type { ExperienceType } from './ExperienceType';
import {
    ExperienceTypeFromJSON,
    ExperienceTypeFromJSONTyped,
    ExperienceTypeToJSON,
    ExperienceTypeToJSONTyped,
} from './ExperienceType';
import type { Producer } from './Producer';
import {
    ProducerFromJSON,
    ProducerFromJSONTyped,
    ProducerToJSON,
    ProducerToJSONTyped,
} from './Producer';
import type { Coordinator } from './Coordinator';
import {
    CoordinatorFromJSON,
    CoordinatorFromJSONTyped,
    CoordinatorToJSON,
    CoordinatorToJSONTyped,
} from './Coordinator';

/**
 * 
 * @export
 * @interface ExperiencesResponse
 */
export interface ExperiencesResponse {
    /**
     * コーディネータ一覧
     * @type {Array<Coordinator>}
     * @memberof ExperiencesResponse
     */
    coordinators: Array<Coordinator>;
    /**
     * 体験種別一覧
     * @type {Array<ExperienceType>}
     * @memberof ExperiencesResponse
     */
    experienceTypes: Array<ExperienceType>;
    /**
     * 体験一覧
     * @type {Array<Experience>}
     * @memberof ExperiencesResponse
     */
    experiences: Array<Experience>;
    /**
     * 生産者一覧
     * @type {Array<Producer>}
     * @memberof ExperiencesResponse
     */
    producers: Array<Producer>;
    /**
     * 体験合計数
     * @type {number}
     * @memberof ExperiencesResponse
     */
    total: number;
}

/**
 * Check if a given object implements the ExperiencesResponse interface.
 */
export function instanceOfExperiencesResponse(value: object): value is ExperiencesResponse {
    if (!('coordinators' in value) || value['coordinators'] === undefined) return false;
    if (!('experienceTypes' in value) || value['experienceTypes'] === undefined) return false;
    if (!('experiences' in value) || value['experiences'] === undefined) return false;
    if (!('producers' in value) || value['producers'] === undefined) return false;
    if (!('total' in value) || value['total'] === undefined) return false;
    return true;
}

export function ExperiencesResponseFromJSON(json: any): ExperiencesResponse {
    return ExperiencesResponseFromJSONTyped(json, false);
}

export function ExperiencesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExperiencesResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'coordinators': ((json['coordinators'] as Array<any>).map(CoordinatorFromJSON)),
        'experienceTypes': ((json['experienceTypes'] as Array<any>).map(ExperienceTypeFromJSON)),
        'experiences': ((json['experiences'] as Array<any>).map(ExperienceFromJSON)),
        'producers': ((json['producers'] as Array<any>).map(ProducerFromJSON)),
        'total': json['total'],
    };
}

export function ExperiencesResponseToJSON(json: any): ExperiencesResponse {
    return ExperiencesResponseToJSONTyped(json, false);
}

export function ExperiencesResponseToJSONTyped(value?: ExperiencesResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'coordinators': ((value['coordinators'] as Array<any>).map(CoordinatorToJSON)),
        'experienceTypes': ((value['experienceTypes'] as Array<any>).map(ExperienceTypeToJSON)),
        'experiences': ((value['experiences'] as Array<any>).map(ExperienceToJSON)),
        'producers': ((value['producers'] as Array<any>).map(ProducerToJSON)),
        'total': value['total'],
    };
}