// web/admin/src/types/api/v1/models/GetUploadURLRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 管理者向け
 * 管理者向けのふるマルAPIです。
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetUploadURLRequest
 */
export interface GetUploadURLRequest {
    /**
     * ファイル種別
     * @type {string}
     * @memberof GetUploadURLRequest
     */
    fileType: string;
}

/**
 * Check if a given object implements the GetUploadURLRequest interface.
 */
export function instanceOfGetUploadURLRequest(value: object): value is GetUploadURLRequest {
    if (!('fileType' in value) || value['fileType'] === undefined) return false;
    return true;
}

export function GetUploadURLRequestFromJSON(json: any): GetUploadURLRequest {
    return GetUploadURLRequestFromJSONTyped(json, false);
}

export function GetUploadURLRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetUploadURLRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'fileType': json['fileType'],
    };
}

export function GetUploadURLRequestToJSON(json: any): GetUploadURLRequest {
    return GetUploadURLRequestToJSONTyped(json, false);
}

export function GetUploadURLRequestToJSONTyped(value?: GetUploadURLRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'fileType': value['fileType'],
    };
}