// web/admin/src/types/api/v1/models/UploadURLResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ふるマル API - 管理者向け
 * 管理者向けのふるマルAPIです。
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UploadURLResponse
 */
export interface UploadURLResponse {
    /**
     * アップロード後の状態参照用キー
     * @type {string}
     * @memberof UploadURLResponse
     */
    key: string;
    /**
     * アップロード用の署名付きURL
     * @type {string}
     * @memberof UploadURLResponse
     */
    url: string;
}

/**
 * Check if a given object implements the UploadURLResponse interface.
 */
export function instanceOfUploadURLResponse(value: object): value is UploadURLResponse {
    if (!('key' in value) || value['key'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    return true;
}

export function UploadURLResponseFromJSON(json: any): UploadURLResponse {
    return UploadURLResponseFromJSONTyped(json, false);
}

export function UploadURLResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UploadURLResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'key': json['key'],
        'url': json['url'],
    };
}

export function UploadURLResponseToJSON(json: any): UploadURLResponse {
    return UploadURLResponseToJSONTyped(json, false);
}

export function UploadURLResponseToJSONTyped(value?: UploadURLResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'key': value['key'],
        'url': value['url'],
    };
}